requirements.txt
boto3==1.36.22
botocore==1.36.22
datasets==3.3.1
matplotlib==3.10.0
networkx==3.4.2
pandas==2.2.3
Requests==2.32.3
scikit_learn==1.6.1
spacy==3.8.4
openai==1.63.2
python-dotenv==1.0.1
SPARQLWrapper==2.0.0
spacy-entity-linker==1.0.3
tqdm